+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.1.1] - 2025-02-11
+### Added
+- return invoice url for order
+- add customer_id for subscription
+
 ## [1.1.0] - 2025-02-07
 ### Added
 - Add parameters discounts, products and cartId list in subscriptions